review fixes from mariusz

+++ libraries/plugins/apis/condenser_api/include/hive/plugins/condenser_api/condenser_api_legacy_operations.hpp
@@ -1403,7 +1403,7 @@ namespace hive { namespace plugins { namespace condenser_api {
     account_name_type to;
     legacy_asset      amount;
     string            memo;
-    uint64_t          recurrence;
+    uint16_t          recurrence;
     time_point_sec    end_date;
   };
